requirements.txt
numpy==1.10.4
Pillow==3.1.1
wordcloud==1.2.1
